Allow more customization of maintenance page file
In the Turnout config, you can specify the following:

public_path - the default is ‘public’ but this allows you to use a sub
directory like ‘public/maintenance_files’ so you can have css and font
files.

maintenance_page_file_name - the default is ‘maintenance’ but this lets
you specify a different file name. It will still use the ‘.html’ or
‘.json’ extension based on what you specify for default_maintenance_page